refactor(angular): rearrange select code
