+++ src/angular/select/public-api.ts
@@ -1,3 +1,3 @@
 export * from './select.module';
-export * from './select/select.component';
+export * from './select/select';
 export * from './select-animations';
