+++ src/angular/select/select.module.ts
@@ -4,7 +4,7 @@ import { NgModule } from '@angular/core';
 import { SbbOptionModule } from '@sbb-esta/angular/core';
 import { SbbIconModule, ɵSBB_ICON_REGISTRY_WRAPPER_PROVIDER } from '@sbb-esta/angular/icon';
 
-import { SbbSelect, SBB_SELECT_SCROLL_STRATEGY_PROVIDER } from './select/select.component';
+import { SbbSelect, SBB_SELECT_SCROLL_STRATEGY_PROVIDER } from './select/select';
 
 @NgModule({
   imports: [SbbIconModule, SbbOptionModule, CommonModule, OverlayModule],
